chore(misunderstood): move ghost mock next to ghost code
