+++ modules/misunderstood/src/backend/db.test.ts
@@ -1,7 +1,7 @@
 import Db, { EVENTS_TABLE_NAME, TABLE_NAME } from './db'
 import { FLAGGED_MESSAGE_STATUS, FLAG_REASON } from '../types'
 
-import { makeMockGhost } from './mockGhost.test'
+import { makeMockGhost } from '../../../../packages/bp/src/core/bpfs/mockGhost.test'
 
 import Database from '../../../../packages/bp/src/core/database'
 import { createDatabaseSuite } from '../../../../packages/bp/src/core/database/index.tests'
